+++ src/components/Layout.js
@@ -13,23 +13,13 @@ const _ = {
   get
 }
 
-//testing
-import { NAVBAR_POSITION_UPON } from '../constants/page-themes'
-import { mockArticle } from '../constants/mock-data'
-
-// Remember to pass following params to header and footer
-// fontColor, pgcolor
-// isIndex, logoColor to header,
-// LogoColor: 'dark' or 'bright'
-
-
 class Layout extends PureComponent {
   render() {
     const pageTheme = _.get(this.props, 'header.pageTheme')
 
     return (
       <div className={classNames(styles.theme, { [styles.photography]: pageTheme === DARK })}>
-        { mockArticle.theme.headerPosition === NAVBAR_POSITION_UPON && this.props.device === 'desktop' ?  null : <Header isIndex={false} pageTheme={pageTheme} pathName={this.props.pathname} headerPosition={mockArticle.theme.headerPosition} /> }
+        <Header isIndex={false} pageTheme={pageTheme} pathName={this.props.pathname}/>
         {this.props.children}
         <Footer pageTheme={pageTheme} />
       </div>
@@ -39,14 +29,12 @@ class Layout extends PureComponent {
 
 Layout.propTypes = {
   header: PropTypes.object,
-  pathname: PropTypes.string.isRequired,
-  device: PropTypes.string
+  pathname: PropTypes.string.isRequired
 }
 
 Layout.defaultProps = {
   header: {},
-  pathname: '',
-  device: ''
+  pathname: ''
 }
 
 export default Layout