+++ src/containers/SherrySpecial.js
@@ -1,62 +1,79 @@
 import { Article, mapStateToProps } from './Article'
-import React, { PureComponent } from 'react'
-import recognizer, { twrSubstring } from '../helpers/url-recognizer'
-import withLayout, { defaultTheme } from '../helpers/with-layout'
-
-import { FadeText } from '@twreporter/react-components'
-import ZoomInImage from '../components/article/zoom-in-leading-image'
 import { connect } from 'react-redux'
+import { FadeText } from '@twreporter/react-components'
+import customizedPathSet from '../constants/customized-article-path'
 import get from 'lodash/get'
+import React from 'react'
 import twreporterRedux from '@twreporter/redux'
+import withLayout, { defaultTheme } from '../helpers/with-layout'
+import ZoomInImage from '../components/article/zoom-in-leading-image'
 
 const _ = {
   get
 }
 
+const setting = {
+  fadeText: {
+    style: {
+      background: 'black',
+      fontColor: 'white'
+    },
+    duration: {
+      textInOut: 5,
+      bgOut: 1
+    },
+    frame: {
+      one: '父母吸毒、家人入獄、貧困、失業、重病……台灣每年至少2萬名少年，在這樣的風險環境裡成長生活。',
+      two: '他們為金錢憂慮、被教育和社會排除，他們很早進入童工之門、放棄學習。',
+      three: '對這群少年而言，活下去，都是一場搏鬥。'
+    }
+  },
+  zoomInImage: {
+    duration: {
+      imageZoomIn: 3,
+      textIn: 1
+    }
+  }
+}
 
-// Fade Text Style
-const BG_COLOR = 'black'
-const FONT_COLOR = 'white'
-// Frame Text
-const FRAME_ONE = '父母吸毒、家人入獄、家庭分崩離析，台灣每年至少2萬名少年成長在這樣的環境。'
-const FRAME_TWO = '他們很早進入童工之門，做著高風險低薪的零工，被迫承擔大人的責任。'
-const FRMAE_THREE = '對這群少年而言，活下去，都是一場搏鬥。'
-// FadeText Duration Setting
-const FadeTextDuration = 5
-const bgOutDuration = 1
-// ZoomInImage Duration Setting
-const imageDuration = 3
-const textDuration = 1
+const fadeText = setting.fadeText
+const zoomInImage = setting.zoomInImage
+const textArr = [ fadeText.frame.one, fadeText.frame.two, fadeText.frame.three ]
 
-const textArr = [ FRAME_ONE, FRAME_TWO, FRMAE_THREE ]
 
 const { actions } = twreporterRedux
 const { fetchAFullPost } = actions
 const IF_DELEGATE_IMAGE = true
 
+const params = { slug: customizedPathSet.HIGH_RISK_YOUTH }
 
-class SherrySpecial extends PureComponent {
+class SherrySpecial extends Article {
   static fetchData(props) {
-    return Article.fetchData(props)
-  }
-
-  constructor(props) {
-    super(props)
+    const { store, pathname } = props
+    const regexp = /[^\/]+(?=\/$|$)/
+    const slug = _.get(pathname.match(regexp), 0, '')
+    const localProps = {
+      store,
+      params: {
+        slug
+      }
+    }
+    return Article.fetchData(localProps)
   }
 
   render() {
-    const { fetchAFullPost, entities, selectedPost, theme, location } = this.props
+    const { entities, theme, location } = this.props
     const originalPathname = _.get(location, 'pathname', '')
-    const pathname = twrSubstring(originalPathname)
-    const params = recognizer({}, pathname)
     return (
       <div>
-        <FadeText
-          bgColor={BG_COLOR}
+        <FadeText.component
+          bgColor={fadeText.style.background}
           textArr={textArr}
-          fontColor={FONT_COLOR}
-          duration={FadeTextDuration}
-          bgOutDuration={bgOutDuration}
+          fontColor={fadeText.style.fontColor}
+          duration={fadeText.duration.textInOut}
+          bgOutDuration={fadeText.duration.bgOut}
+          ifLock={true}
+          ifUnlock={false}
         />
         {
           IF_DELEGATE_IMAGE ?
@@ -65,37 +82,38 @@ class SherrySpecial extends PureComponent {
             entities={entities}
             params={params}
             pathname={originalPathname}
-            totalDelay={FadeTextDuration * textArr.length + bgOutDuration}
-            imageDuration={imageDuration}
-            textDuration={textDuration}
+            totalDelay={fadeText.duration.textInOut * textArr.length + fadeText.duration.bgOut}
+            imageDuration={zoomInImage.duration.imageZommIn}
+            textDuration={zoomInImage.duration.textIn}
           /> : null
         }
-        <Article
-          fetchAFullPost={fetchAFullPost}
-          entities={entities}
-          selectedPost={selectedPost}
-          theme={theme}
-          params={params}
-          ifDelegateImage={IF_DELEGATE_IMAGE}
-        />
+        {super.render()}
       </div>
     )
   }
 }
 
+const enhancedMapStateToProps = (state) => {
+  const superProps = mapStateToProps(state)
+  superProps.ifDelegateImage = IF_DELEGATE_IMAGE
+  superProps.params = params
+  return superProps
+}
 
 SherrySpecial.propTypes = {
   entities: React.PropTypes.object,
   selectedPost: React.PropTypes.object,
   theme: React.PropTypes.object,
-  params: React.PropTypes.object
+  params: React.PropTypes.object,
+  ifDelegateImage: React.PropTypes.bool
 }
 
 SherrySpecial.defaultProps = {
   entities: {},
   selectedPost: {},
   theme: defaultTheme,
-  params: {}
+  params: {},
+  ifDelegateImage: false
 }
 
-export default connect(mapStateToProps, { fetchAFullPost })(withLayout(SherrySpecial))
+export default connect(enhancedMapStateToProps, { fetchAFullPost })(withLayout(SherrySpecial))