+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+### Unreleased
+### 2.2.4
+- @twreporter/react-components version number to 1.1.4
 ### 2.2.3
+- Add center-small alignment style in src/components/article/BlockAlignmentWrapper.scss.
 - Update article to be compatible with dynamic themes
 - Create customized article.
 - Create zoom in leading image.
+- Fix scroll to top problem.
+- @twreporter/react-components version number to 1.1.3
+- Add TopicLandingPage topic description iframe max-width 100% for youtube player or other embedded items.
 
 ### 2.2.2
 - Use npm-scope. Update twreporter-react-components to @twreporter/react-components and twreporter-redux to @twreporter/redux [#673](https://github.com/twreporter/twreporter-react/pull/673)